+++ src/data_models/WeekYearData.py
@@ -1,5 +1,9 @@
 from dataclasses import dataclass
-from sqlalchemy import Column, Integer
+from sqlalchemy import (
+    Column,
+    Integer,
+    String
+)
 
 from  src.constants import Base
 
@@ -8,5 +12,6 @@ from  src.constants import Base
 class WeekYearData(Base):
     __tablename__ = 'week_year'
     __table_args__ = {'extend_existing': True}
+    id = Column(String(50), primary_key=True)
     week = Column(Integer)
-    year = Column(Integer, primary_key=True)
+    year = Column(Integer)
